v3: Ignore UpdaterBranchSelected
This event continued to be sent in eos4 and newer, even though it was
not handled in the v3 event processor.

It will no longer be sent by Endless OS 5.1.3 and 6, spiritually
superseded by the CheckpointBlocked{Daily,Monthly} aggregate events.
Discard the 3.4 million instances of the old event in
unknown_singular_event_v3 and any future submissions.

https://phabricator.endlessm.com/T33684