v3: Ignore ControlCenterAutomaticUpdates
This has not been sent since eos5.0; and when sent by eos4.0 it was not
handled in the v3 event processor.

For posterity, based on the ~40,000 rows in the
unknown_singular_event_v3 table: this event was sent ~2,000 times per
month during the time when eos4.0 was the latest release. I haven't
analysed the contents.

https://phabricator.endlessm.com/T33684