v2: Discard Uptime table
This table is very large and the data is not useful to us.  Drop the
table; and drop any future instances of this event that are received.

This table was used in the test suite not just for its own sake, but
also to exercise the handling of the __ignore_empty_payload__ flag on
models compared to non-empty invalid payloads when replaying events.

The only other v2 event that sets __ignore_empty_payload__ to True is
LocationLabel, which has special handling to upsert the label into the
machine table, so isn't trivial to swap in. Instead write a dedicated
test alongside the other tests of the special handling of LocationLabel.

Even so, this patch renders some lines of the implementation of
replaying unknown and invalid events newly uncovered by tests. But we
will only be deleting from the v2 event processor, not adding to it, so
I am not too worried about that.

https://phabricator.endlessm.com/T33684