Allow events to accept payloads of multiple types
This will be useful to add an additional field to the ComputerInformation
event.

The more principled alternative would be to allow events to match multiple
UUIDs with a different UUID for each payload type. However this would be a much
more invasive refactoring.

It's quite annoying to have these rigid types for this event, where the input
is (derived from) a JSON document (namely `lscpu --json`) and the storage is
also a jsonb column. Oh well.

https://phabricator.endlessm.com/T35351