endless: Handle image ID "[Invalid UTF-8]" like "unknown"
Endless OS 4.0.x and 5.0.x would submit this string if the image version could
not be read as valid UTF-8 from disk. (This string comes from within GLib if
you pass a NULL pointer as the argument for a 's' type code in a call to
g_variant_new().)

This is fixed in Endless OS 5.1.0 and newer, but we have many requests in the
error queue which are failing to be loaded for this reason.

Treat "[Invalid UTF-8]" the same as "unknown".

https://phabricator.endlessm.com/T35279


https://github.com/endlessm/eos-event-recorder-daemon/commit/13d533b5d47cc592545cbcb63cb17c28df860762