tests: Rework temporary handlers
The azafea configuration expects that handlers are packages with
migration directories within them. Several integration tests using their
own handlers weren't following that convention and fail with newer
alembic.

In order to make them work in that fashion, make a temporary package
with the handler module code copied into it. Since the migrations are in
a temporary directory, all of that cleanup goes away. For
`IntegrationTest`, allow setting a handler module file path to trigger
the temporary handler package mode. Tests that need to import the
handler module need to do so from the copied path so that sqlalchemy
doesn't see duplicated tables.

While here, add some logging to `IntegrationTest` since it does a lot of
very non-trivial things and the messages make it easier to pinpoint
where things are going wrong.

This also uses `monkeypatch.syspathprepend` for tweaking `sys.path` and
the `tmp_path` fixture rather than `tmpdir`. `syspathprepend` allows a
`FIXME` to go away as it has a `pkg_resources` workaround. `tmp_path` is
basically the same as `tmpdir`, but the former uses a regular
`pathlib.Path` and the latter is deprecated.