github: Use python container for metabase docs
Now that pipenv specifically wants to use python 3.11, we can't just run
whatever python is in Ubuntu. Instead, use the same base python 3.11
Docker image and repeat the dependency installation that's normally done
in the real image. It would be nice to use a dedicated Dockerfile to
capture this, but GitHub Actions doesn't let you do that for a job
container.

https://phabricator.endlessm.com/T32367