docs: Properly initialize metabase builder
Instead of using class attributes and populating them lazily, add an
`init` method to the builder as that's the proper way to initialize a
Builder subclass. With that, the requests session can be created
immediately since we know it will be needed.

Without initializing the tables dict, it fails when there are no new
docstrings to process since the tables dict was previously initialized
lazily in `process_docstring`.

The worker attribute wasn't actually being used.

https://phabricator.endlessm.com/T32367