Use pipenv sync for installing packages
In modern pipenv, the way to just install packages is to run `sync`.
Apparently, now when you run `install` and it detects that the lock file
is out of date, it will helpfully update the lock file for you. That's
not what we want for our builds. For the docker image, also pass
`--clear` to clear the cached packages.

https://phabricator.endlessm.com/T35141